Create TS0601_rcbo_WDYK_ZJSBL7-100Z_3P.js
This is a ready-made zigbee2mqtt converter for a three-phase energy meter WDYK_ZJSBL7-100Z. Created by extending the standard Taya TS0601_rcbo for Home Assistant 2022.4.5. It is enough to put the file in the zigbee2mqtt directory and register the file name in Home Assistant - zigbee2mqtt - external converters.